requirements.txt
# Core Flask and API dependencies
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.0.5
flask-mail==0.9.1

# Database and authentication
passlib[bcrypt]==1.7.4
mysqlclient==2.1.1
python-dotenv==1.0.0

# Web scraping and automation
selenium==4.15.2
undetected-chromedriver==3.5.4
requests==2.31.0
aiohttp==3.8.3

# AI and content generation
openai==1.3.7
openrouter==1.0

# Reddit API
praw==7.7.1

# Utilities
python-dateutil==2.8.2

# Payment and billing
stripe==7.8.0